refactor(room): the wire adapters get their own module
server.ts was one module edited for two unrelated reasons: room
semantics (membership, validation, relay decisions) and wire adaptation
(peer frames, acks, pre-peer buffering). The two adapters at that seam
— RoomStubChannel and bindParticipantStubChannel — now live in
room/stubs.ts, which imports the domain; the domain only knows the stub
as a type (no runtime cycle). server.ts reads top-to-bottom as one
story again: statics → ServerRoom → participant → store access.

Riders surfaced by the pass:
- bindParticipantStubChannel reached three objects deep
  (participant._room._state.getRemote) — now uses the room's public
  getParticipant(), the same interface every other caller crosses.
- _setMemberMeta rebuilt the member record field-by-field, silently
  dropping any field it doesn't know about — it now spreads the stored
  record, so future record fields survive a setMeta().
- makeEid and errorMessage move to shared.ts: both are protocol
  concerns (echo-dedupe IDs, how errors ride ok:false acks), and both
  sides of the stubs seam need them.
